# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GuimaAPI",
        "description": "API para gerenciamento de alunos, treino, dieta, aulas e avaliações físicas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/modulos": {
            "get": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Listar todos os modulos",
                "operationId": "db59cbb33e25a3bd02924aeb82476456",
                "responses": {
                    "200": {
                        "description": "Retorna a lista de modulos"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Criar um novo modulo",
                "operationId": "e7866d9463e35c03ff375124542e73fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Treinamento"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Modulo de treinamento"
                                    },
                                    "ordem": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "img_reference": {
                                        "type": "string",
                                        "example": "null"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o modulo criado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/modulos/{modulo_id}": {
            "get": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Mostrar um modulo específico por id",
                "operationId": "72a0c0435ced86bff2d94edf8bd1a923",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o modulo específico por id"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Deletar um modulo específico por id",
                "operationId": "4cdb8fd5b0c0b50784e26f2160a177cb",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o modulo deletado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Modulos"
                ],
                "summary": "Atualizar um modulo específico por id",
                "operationId": "6cceaaef24235c451f7c3e05dd26249c",
                "parameters": [
                    {
                        "name": "modulo_id",
                        "in": "path",
                        "description": "Id do modulo",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Treinamento"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Modulo de treinamento"
                                    },
                                    "ordem": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "img_reference": {
                                        "type": "string",
                                        "example": "null"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o modulo atualizado por id"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}